refactor: move monika flag related code into one file to make it easier to change
