+++ src/commands/monika.ts
@@ -35,14 +35,18 @@ import { sendMonikaStartMessage } from '../components/notification/start-message
 import { scheduleSummaryNotification } from '../components/notification/schedule-notification'
 import { setContext } from '../context'
 import events from '../events'
-import { retryInitialDelayMs, retryMaxDelayMs, symonAPIVersion } from '../flag'
+import {
+  type MonikaFlags,
+  monikaFlagsDefaultValue,
+  retryInitialDelayMs,
+  retryMaxDelayMs,
+  symonAPIVersion,
+} from '../flag'
 import type { Config } from '../interfaces/config'
 import type { Probe } from '../interfaces/probe'
 import { printSummary, savePidFile } from '../jobs/summary-notification'
 import initLoaders from '../loaders'
 import { sanitizeProbe, startProbing } from '../looper'
-import { monikaFlagsDefaultValue } from '../context/monika-flags'
-import type { MonikaFlags } from '../context/monika-flags'
 import SymonClient from '../symon'
 import { getEventEmitter } from '../utils/events'
 import { log } from '../utils/pino'
