+++ src/context/monika-flags.ts
@@ -24,68 +24,8 @@
 
 import fs from 'fs'
 import path from 'path'
-import { SYMON_API_VERSION } from '../flag'
 
-export type MonikaFlags = {
-  'auto-update'?: string
-  config: string[]
-  'config-filename': string
-  'config-interval': number
-  'create-config': boolean
-  flush: boolean
-  'follow-redirects': number
-  force: boolean
-  har?: string
-  id?: string
-  insomnia?: string
-  json?: boolean
-  'keep-verbose-logs': boolean
-  logs: boolean
-  'one-probe': boolean
-  output?: string
-  postman?: string
-  prometheus?: number
-  repeat: number
-  retryInitialDelayMs: number
-  retryMaxDelayMs: number
-  sitemap?: string
-  'status-notification'?: string
-  stun: number
-  summary: boolean
-  symonKey?: string
-  symonLocationId?: string
-  symonMonikaId?: string
-  symonReportInterval?: number
-  symonReportLimit?: number
-  symonUrl?: string
-  'symon-api-version'?: SYMON_API_VERSION
-  text?: string
-  verbose: boolean
-  version: void
-}
-
-export const monikaFlagsDefaultValue: Pick<
-  MonikaFlags,
-  | 'config'
-  | 'config-filename'
-  | 'config-interval'
-  | 'follow-redirects'
-  | 'retryInitialDelayMs'
-  | 'retryMaxDelayMs'
-  | 'stun'
-  | 'symon-api-version'
-> = {
-  config: getDefaultConfig(),
-  'config-filename': 'monika.yml',
-  'config-interval': 900,
-  'follow-redirects': 21, // axios default https://github.com/axios/axios
-  retryInitialDelayMs: 128,
-  retryMaxDelayMs: 30_000,
-  // default is 20s interval lookup
-  stun: 20,
-}
-
-function getDefaultConfig(): Array<string> {
+export function getDefaultConfig(): Array<string> {
   const filesArray = fs.readdirSync(path.dirname('../'))
   const monikaDotJsonFile = filesArray.find((x) => x === 'monika.json')
   const monikaDotYamlFile = filesArray.find(
