+++ src/flag.ts
@@ -24,16 +24,77 @@
 
 import { Flags } from '@oclif/core'
 
+import { getDefaultConfig } from './context/monika-flags'
+
 export enum SYMON_API_VERSION {
   'v1' = 'v1',
   'v2' = 'v2',
 }
 
+export type MonikaFlags = {
+  'auto-update'?: string
+  config: string[]
+  'config-filename': string
+  'config-interval': number
+  'create-config': boolean
+  flush: boolean
+  'follow-redirects': number
+  force: boolean
+  har?: string
+  id?: string
+  insomnia?: string
+  json?: boolean
+  'keep-verbose-logs': boolean
+  logs: boolean
+  'one-probe': boolean
+  output?: string
+  postman?: string
+  prometheus?: number
+  repeat: number
+  retryInitialDelayMs: number
+  retryMaxDelayMs: number
+  sitemap?: string
+  'status-notification'?: string
+  stun: number
+  summary: boolean
+  'symon-api-version'?: SYMON_API_VERSION
+  symonKey?: string
+  symonLocationId?: string
+  symonMonikaId?: string
+  symonReportInterval?: number
+  symonReportLimit?: number
+  symonUrl?: string
+  text?: string
+  verbose: boolean
+  version: void
+}
+
+export const monikaFlagsDefaultValue: MonikaFlags = {
+  config: getDefaultConfig(),
+  'config-filename': 'monika.yml',
+  'config-interval': 900,
+  'create-config': false,
+  flush: false,
+  'follow-redirects': 21,
+  force: false,
+  'keep-verbose-logs': false,
+  logs: false,
+  'one-probe': false,
+  repeat: 0,
+  retryInitialDelayMs: 128,
+  retryMaxDelayMs: 30_000,
+  // default is 20s interval lookup
+  stun: 20,
+  summary: false,
+  verbose: false,
+  version: undefined,
+}
+
 export const symonAPIVersion = Flags.custom<SYMON_API_VERSION>({
   default: SYMON_API_VERSION.v1,
-  options: [SYMON_API_VERSION.v1, SYMON_API_VERSION.v2],
   description:
     'Symon API version to use. Available options: v1, v2. Default: v1',
+  options: [SYMON_API_VERSION.v1, SYMON_API_VERSION.v2],
 })
 
 export const retryInitialDelayMs = Flags.integer({
