+++ src/loaders/index.ts
@@ -26,7 +26,7 @@ import type { Config as IConfig } from '@oclif/core'
 import { isSymonModeFrom, setupConfig } from '../components/config'
 import { setContext } from '../context'
 import events from '../events'
-import type { MonikaFlags } from '../context/monika-flags'
+import type { MonikaFlags } from '../flag'
 import { tlsChecker } from '../jobs/tls-check'
 import type { Probe } from '../interfaces/probe'
 import { loopCheckSTUNServer } from '../looper'
