+++ src/symon/index.ts
@@ -34,7 +34,7 @@ import path from 'path'
 import { updateConfig } from '../components/config'
 import { getOSName } from '../components/notification/alert-message'
 import { getContext } from '../context'
-import type { MonikaFlags } from '../context/monika-flags'
+import type { MonikaFlags } from '../flag'
 import events from '../events'
 import { Config } from '../interfaces/config'
 import { Probe } from '../interfaces/probe'
